+++ internal/pve_utils/host.go
@@ -1,0 +1,17 @@
+package pveutils
+
+import "os/exec"
+
+func IsPVEHost() bool {
+	cmd := exec.Command("/usr/bin/which", "pvedaemon")
+	if cmd == nil {
+		return false
+	}
+
+	_, err := cmd.Output()
+	if err != nil {
+		return false
+	}
+
+	return true
+}