+++ pkg/helpers/regexp.go
@@ -1,0 +1,19 @@
+package helpers
+
+import "regexp"
+
+// GetRegexpParams parses url with the given regular
+// expression and returns the group values defined
+// in the expression.
+func GetRegexpParams(regEx, url string) (paramsMap map[string]string) {
+	var compRegEx = regexp.MustCompile(regEx)
+	match := compRegEx.FindStringSubmatch(url)
+
+	paramsMap = make(map[string]string)
+	for i, name := range compRegEx.SubexpNames() {
+		if i > 0 && i <= len(match) {
+			paramsMap[name] = match[i]
+		}
+	}
+	return paramsMap
+}